chore: relocate `getLocation` utility to `location.ts` for better modularity
